Sillly bug fixed all CRUD works now

+++ app/Http/Controllers/UserController.php
@@ -6,6 +6,7 @@ use App\Models\User;
 use App\Http\Requests\StoreUserRequest;
 use App\Http\Requests\UpdateUserRequest;
 use App\Models\Role;
+use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
 {
@@ -16,7 +17,9 @@ class UserController extends Controller
      */
     public function index()
     {
-        $users = User::with('roles')->get();
+        $this->authorize('viewAny', User::class);
+
+        $users = User::with('roles')->simplePaginate(10);
         return view('webpages.users.index', compact('users'));
     }
 
@@ -27,8 +30,9 @@ class UserController extends Controller
      */
     public function create()
     {
-        $roles = Role::pluck('title', 'id');
+        $this->authorize('create', User::class);
 
+        $roles = Role::pluck('title', 'id');
         return view('webpages.users.create', compact('roles'));
     }
 
@@ -41,11 +45,11 @@ class UserController extends Controller
     public function store(StoreUserRequest $request)
     {
         $user = User::create($request->validated());
+
         $user->fill([
-            'password' => bcrypt($request->newPassword)
+            'password' => Hash::make($request->input('password'))
         ])->save();
         $user->roles()->sync($request->input('roles', []));
-
         return redirect()->route('users.index');
     }
 
@@ -57,6 +61,8 @@ class UserController extends Controller
      */
     public function show(User $user)
     {
+        $this->authorize('view', $user);
+
         return view('webpages.users.show', compact('user'));
     }
 
@@ -68,6 +74,8 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
+        $this->authorize('update', $user);
+
         $roles = Role::pluck('title', 'id');
         $user->load('roles');
 
@@ -83,11 +91,10 @@ class UserController extends Controller
      */
     public function update(UpdateUserRequest $request, User $user)
     {
-
-
         $user->update($request->validated());
+
         $user->fill([
-            'password' => bcrypt($request->newPassword)
+            'password' => Hash::make($request->input('password'))
         ])->save();
         $user->roles()->sync($request->input('roles', []));
 
@@ -102,8 +109,9 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
-        $user->delete();
+        $this->authorize('delete', $user);
 
+        $user->delete();
         return redirect()->route('users.index');
     }
 }
